Change internal representation of events
Move to a representation that more closely resembles the notation of the
article. For example, loss events do not store the synteny as it is
after the loss has occurred, but rather as it is before, and also store
the interval (segment) of genes that are lost. Furthermore, the
duplicated segment is now stored alongwith the duplication event by
itself.

Refactor to merge the two functions that generate losses and duplication
segments as a single one that only emits a segment and does not change
the parent synteny whatsoever.

This change of representation enables the ./viz program to visualize
trees that contain rearrangements, because it does not need to compute
the difference between a loss node and its parent to infer back the lost
segment (this computation previously required that the loss node was in
the same order as its parent, which is not verified when rearrangements
occur.)